Add a "pass-cs-fd" service option, which passes a control socket file descriptor directly to service processes, allowing them issue commands even if the regular control socket has not been created (or if they will not have permission to access it).
This has limited use cases and has security implications when used,
but allows early launching of a service which can pass events / issue
commands to dinit. In particular this could be useful for a device
watcher daemon which starts services as certain devices come online
(eg check and mount disks as they are detected by the kernel).