Inhibit restart of dependents when service manually stopped
Previously, manually stopping a service inhibited automatic restart of
the service itself, but not of any dependents. As a result the service
could still restart (because a dependent would restart and require the
service). Since this is somewhat inconsistent, change the behaviour so
that manually stopping a service inhibits restarts of dependents also
(and thus completely inhibits restart of the specified service).