Improve control socket creation logic.
Attempt to check for existing, active socket before unlinking it, and
refuse to start if there is one. Unfortunately this check cannot be done
atomically.

If the socket doesn't exist or appears stale, unlink it even if we
aren't system init. This will most likely give a better experience for
users.