Default to .config for user services directory
The /etc directory is a directory for system configuration files.
The .config directory in most Linux systems nowadays mirrors this.

Explicitly do not follow XDG_CONFIG_HOME as that is of questionable
value and the XDG base directory spec is overly complicated and
has various quirky parts. However, IME most users appreciate having
config files in a separate location and .config is where a lot of
software stores these nowadays.