Initial support for cgroups
This is mainly in the form of a "run-in-cgroup" setting where a cgroup
name (path) can be specified (a relative path will be resolved relative
to the cgroup of dinit).

Only a single hierarchy is supported. Cgroups v2 are a natural fit (and
have been tested) but Cgroups v1 with a single hierarchy should probably
also work.

There is no functionality to create cgroups or set their resource
constraints; that needs to be done separately.