Fix #74, command arg ptrs pointed into invalid string
Command line arguments were constructed as an array of pointers into a
string, but the string was then moved into the service record. This
rarely manifested an issue since a string beyond a short length will
allocate its data on the heap and moving the string will not change the
heap allocation, but for short strings this is not always the case. In
any case, you are not supposed to retain pointers to the string
characters after moving the string.

Resolve this by adding a "ha_string" class which always heap-allocates
for the string data, and which explicitly always retains the heap
allocation on move.

Reported by Duncan Overbruck (Duncaen).