Always re-attach dependency acquisitions if service states are amenable
I.e. if a dependency is added between two started services (or starting
sevices) add an acquisition. This means the dependent will keep the
dependency started in cases where it would otherwise be stopped due to
having no active dependents.

The should mean that acquisition is now invariant, i.e. if two services
are in the appropriate states then there should be a an acquisition.
This is consistent with starting a dependent service where the dependent
was already started; in that case the acquisition is (already)
re-attached.