Allow restart of services if all dependents restart automatically (#108)
Allow restarting services with hard dependents (with --force)

If the user requested restart involves a pinned started dependency, the
restart will not be allowed. If any hard dependents exist, --force is
required.

If the restart is allowed then the restart is propagated to all hard
dependents and their hard dependents.