write_all[_x]: change count parameter type to int
write(2) accepts size_t, so this change artificially limits the amount
that be written by write_all(), but we only use it for small buffers so
that's fine (and means we don't need to increase the precision of the
return type).

Also return the count of bytes written correctly in case of an error
after a partial write.