Fix for not correctly buffering subprocess output in shutdown
We were closing the output pipe *before* forking instead of just after.
As a result, the subprocesses (eg umount, swapoff) would have put output
directly to stdout. In most cases this is harmless, however, the reason
for the buffering is so that shutdown can execute even if stdout (which
may be a serial line etc) becomes blocked.