make: pass SANITIZEOPTS last
As things are, user-passed flags override the sanitizer options
that are used for tests. The UBSan sanitizer has a minimal runtime
option, which can be used in production due to minimal overhead
while still printing reasonable messages. However, this minimal
runtime is not compatible with ASan, which is used for tests by
default, therefore it is necessary to disable this minimal
runtime for those files.

Unless SANITIZEOPTS are passed last, the potential minimal-runtime
option passed from flags used for the rest of the build may result
in failed test builds when SANITIZEOPTS contain ASan. Even if
passed last, one still gotta disable the minimal runtime inside
of SANITIZEOPTS, but that is something the user can control.