Add kill-all-on-stop service option and documentation
"Before stopping this service, send a TERM signal and then (after a
short pause) a KILL signal to all other processes in the system,
forcibly terminating them. This option is intended to allow system
shutdown scripts to run without any possible interference from
"leftover" or orphaned processes (for example, unmounting file systems
usually requires that the file systems are no longer in use).

This option must be used with care since the signal broadcast does not
discriminate and potentially kills other services (or their shutdown
scripts); a strict dependency ordering is suggested, i.e. every other
service should either be a (possibly transitive) dependency or dependent
of the service with this option set.

This option can be used for scripted and internal services only."