Restore behaviour of checking for control socket file and re-creating it
Dinit 0.16.x, after starting a "starts-rwfs"-marked service, will check
whether the control socket file still exists even if it has already
managed to create the control socket. If it doesn't exist the socket
will be closed and then re-created.

This behavior was removed since, but since this may lead to surprise
boot failures if services are not configured correctly, I am restoring
the behaviour.

See #148