Implement output capture to buffer
The new 'log-type = buffer' setting can be used to specify that output
should be captured to a buffer in memory rather than to file. The
maximum size (in bytes) can be specified via 'log-buffer-size'.

"dinitctl catlog <service-name>" can be used to view the captured output
from a service.

Documentation pending.