Rename HOSTxxx make variables to xxx_FOR_BUILD
As pointed out by Eli Schwartz, autotools uses "host" (from
build/host/target) to mean "machine the compiled program will run on"
rather than "the machine that performs the build". Thus HOSTCXX could be
mistaken for meaning the opposite of what it does.

So, use CXX_FOR_BUILD instead of HOSTCXX (for example). It's not an
ideal name, but it's what autotools uses, and I can't come up with
anything better.