dinitctl: add is-active and is-failed commands
The is-active command makes dinitctl return 0 when the given
service is currently active, and 1 in any other scenario. The
is-failed command makes it return 0 when the given service has
failed to start and is currently stopped, and 1 in any other
scenario. The service not existing and so on counts as non-success
for both commands.

The purpose of these is to use them in scripts, as parsing status
output is clunky. In general it is enough to pass --quiet and just
check the status code, but both commands also print the current
service status (and nothing else) if in default verbose mode.