Define USE_INITGROUPS unconditionally via mconfig-gen
Default to 1 (do use initgroups) if unset. POSIX apparently refuses any
responsibility for functions which set supplementary groups, but they
are fundamentally necessary and so generally supported anyway.

Since we now define it unconditionally in mconfig.h, there's no need to
add a default case in run-child-proc.cc, so also remove that.