igr_functions: make use of IGR_OUTPUT consistent.
The way it was written, if IGR_OUTPUT was set on input it was modified
by appending the test name, if it wasn't set it was set to a directory
in the current working directory. This double use was strange and
confusing.

Now use IGR_OUTPUT_BASE as the base directory for output, if it is set
(IGR_OUTPUT is set to $IGR_OUTPUT_BASE/$TEST_NAME). Otherwise IGR_OUTPUT
is set to $PWD/output. It is still somewhat different depending on
whether IGR_OUTPUT_BASE is set or not, but at least use of IGR_OUTPUT
itself is consistent.