logfile_gid: skip unnecessary getgid() call
The logic for setting logfile_gid in settings wasn't really correct,
though it gave the correct result anyway. logfile_uid_gid, supposedly
representing the primary group of the user specified in logfile_uid, was
being set via getgid() which is incorrect, and logfile_gid was then set
to this if it was unknown/unset.

By switching that to set logfile_gid to logfile_uid_gid first, and then
setting it via getgid() if still unknown, we avoid an unnecessary
getgid() call and avoid setting logfile_uid_gid to the wrong value (even
though it is not used later anyway, so this isn't terribly important).