Fix behaviour of group initialisation when run-as specified by uid
The "run-as" setting can be specified via numerical user id rather than
username. This is intended mainly for cases where services may be loaded
before the system user ("passwd") and group database are available, eg.
in early initial ramdisk.

Documentation has always stated that if a numerical setting is used, the
group is not looked up from the database and the process runs as the
same group as dinit. The behaviour had diverged from this; restore that
behaviour now. In addition, clarify in the documentation that
supplementary groups are dropped in this case.