Improve signal handling for service processes
Keep the original signal mask that is set when dinit starts and apply it
(minus a few specific signals) to service processes. Mask all signals
that can be invoked via the terminal by default; add an option,
"unmask-intr", to unmask SIGINT.

In addition, mask all signals after starting if dinit is running in
system manager mode, to avoid the risk of death of suspension by an
unhandled signal.