handle_exit_status: remove dead stores to service_status
The service_status variable was actually a local variable, shadowing the
field (which is anyway private in service_record). Setting it wasn't
doing anything (and we don't need to set the state before calling
failed_to_start(), since that sets the state to STOPPED anyway).

Rename the variable now as well to prevent future confusion.