Make a better effort to use correct target for symlink on enable
Enabling a service creates a symbolic link in the waits-for.d direcetory
specified in the "from" service (usually boot).

The link target doesn't actually matter to Dinit. Previously we just
used "../to-service-name" (where "to-service-name" was the name of the
service that was being enabled) but this is not always correct.

Make a "good" effort to figure out a suitable link target. However, we
still don't (and won't) try to get it 100% correct; it's not worth the
effort of dealing with resolving relative paths, canonicalising them,
and generating a relative path.

Relates to #129