dinit-monitor: re-word and fix formatting in documentation
Separate --str-failed/--str-stopped/--str-started documentation into one
paragraph for each.

Fix formatting so that each "--str-xxx" shows in bold in the man page.

Fix formatting in man page so that argument for "--str-xxx" is shown in
"italics" (\fIargument\fR) and not as <argument>. The latter isn't used
in man pages.

Fix "substituion" typo and reword. Use "text" in place of
programmer-centric "string" (though we leave the argument names as is).