dinitctl: slight improvement to err msg when no instance running
It seems like it's common enough for users to be confused when run a
dinitctl command with no user instance, so slightly improve the message
(to suggest the reason for socket connection failure might be the
absence of a user instance). It might not be enough to resolve all
confusion but it is at least an improvement.