Begin re-work of integration tests
The integration test shell scripts are fragile. This change moves
towards running integration tests via the igr-runner directly. A bunch
of utility classes and functions are implemented in igr.h. These allow:

- running dinit and capturing stdout/stderr
- waiting for termination, with a timeout
- checking file contents match a specific string

More will be added soon, as the rest of the integration tests are
converted to use the new framework.

The old shell-script based tests remain in place for now as they are
used in the meson build. However, they will be removed once the meson
build is fixed to use igr-runner.