Add depends-on.d and depends-ms.d options as per waits-for.d
These allow specifying directories which contain files whose names
represent regular (depends-on) or milestone (depends-ms) dependencies.

These (as with waits-for.d) are intended to be useful to distributions,
as they allow a package to easily "add itself" as a dependency on
another simply via the presence of a file (as long as the original
service is appropriately configured).