control: Don't call process_packet() when bad_conn_closed is set
Sending garbage data and then closing the connection could lead into out-of-memory.
Because after closing the connection, Dinit will stuck in (rbuf.get_length() >= chklen)
while loop and then the outbuf buffer is used because the connection fd is not available
anymore. This triggers the OOM detector in the fuzzer and can result in segfault when Dinit
is used as init of system.

Firstly detected in https://github.com/davmac314/dinit/actions/runs/7233843212

Signed-off-by: Mobin Aydinfar <mobin@mobintestserver.ir>