cptests: New test about passing invalid data to control protocol
For early detection of related issues.
See https://github.com/davmac314/dinit/pull/354 for an example.

Also in this commit, setting and getting current watcher flags is implemented.
This is useful for checking current watcher flags to make sure the watcher
is in desired state (Used in this new test).

Signed-off-by: Mobin Aydinfar <mobin@mobintestserver.ir>