Support for "+=" in service settings ("command" only)
This adds support for "+=" operator as a way to append to setting values
in service description files (instead of "=" which is currently used for
assignment). For the moment, it is implemented only for the "command"
setting.

This should be useful for user customisation of system services - the
user can override the system service description, but @include the
original, and then add extra arguments to the command line via "command
+= ...".

Support for additional commands, tests, and documentation will follow.