Unconditionally add whitespace separator when appending setting values
Partly this is to allow for existing components which have white space,
so checking for a trailing white space doesn't really indicate that
appending other text won't be jammed up against the previous value. It
doesn't matter all that much though, since it's really the
part_positions list which maintains the positions of the component
parts, so might as well just make the code simpler.