selinux_transition: always error exit if we fail to load the policy
Returning true if we fail to load the policy while the system is in
permissive mode breaks one of the assumptions of permissive mode: that
new files are labelled correctly according to the system policy (unlike
when booting with all SELinux related infrastructure disabled by passing
selinux=0 to the kernel cmdline). Let's always exit if we we fail to
load the policy; if the loading of the fails, something has gone badly
wrong, and the system should be booted with selinux=0 to disable all
selinux related infrastructure.