selinux_transition: clarify policy choice for inital domain in comment
It's misleading to say that a policy choice is a misconfiguration, the
choice to prevent dinit from transitioning from its inital domain is a
choice made the local system administrator. Let's clarify our comments
to specify that this isn't an error, but a warning, as this is still a
unsual policy choice.