selinux_transition: only bail early if enforcing mode requested
As can be seen in selinux_init_load_policy(3), if the SELinux kernel
infrastructure was requested to not be loaded, the enforce int passed
to the function is set to 0, indicating a non-enforcing load [1]. In the
event of a system where the SELinux policy itself is somehow damaged,
missing or malformed, or any other similar events where the policy is
unable to be loaded, it should be expected by an end user that passing
the selinux=0 flag will disable all of the system's SELinux related
infrastructure.

It should be noted however that this is a compromise. libselinux makes
no distinction between selinux being disabled and permissive mode being
requested in selinux_init_load_policy(3) [2], so in the event of dinit
being unable to load the SELinux policy when requested to load in
permissive mode, we will continue to a boot to a state where the system
will likely end up needing a relabel, as the policy is not loaded and
inplace to create files with correct contexts.

[1] https://github.com/SELinuxProject/selinux/blob/main/libselinux/src/load_policy.c#L299-L306
[2] https://github.com/SELinuxProject/selinux/blob/main/libselinux/src/load_policy.c#L264-L273

Signed-off-by: Rahul Sandhu <rahul@sandhuservices.dev>