Update contribution guidelines
Experience over the past few months/years has shown that a lot of
contributors don't really understand what is expected of them or how the
contribution process is supposed to work.

Try to address that by updating the CONTRIBUTING and DESIGN documents,
and adding a new PR-PROCESS document (Pull Review process).