parse_gid_param: switch arg order of service/parameter name
Other parsing functions in load-service generally have the service name
as an argument with the parameter name as the next argument. For some
reason this was switched around with parse_gid_param.

This already caused one bug, which is also fixed with this commit
("logfile-gid" parameter had it around the wrong way, which would have
caused funky error messages in case of an invalid setting).