meson: Use static_library for shared sources
In Make-based builds tests are build twice with all sources because of
different flags for tests. In meson this is not the case so just link to
shared source instead of rebuiling over and over. Also this fixes building
every test source for every single test target.

This reduces ninja targets count from 98 to just 60.