open_with_dir: function to open a file and its parent directory together
Since we want to resolve references to file paths within service
descriptions (or fragments thereof) relative to the service description
file itself, we need to determine the parent directory of the service
description file. This is tricky in the presence of symbolic links;
since the final path component can be a link, just stripping that final
component doesn't give the real parent directory - we have to fully
resolve the link.

However, the naive approach of realpath'ing the link and then stripping
the final component to find the parent is prone to a race (where any
component of that realpath gets changed to a symlink in between the
realpath operation and opening the parent). It's probably not a security
issue given our security model, but it's still opening up the small
possibility of file paths being resolve incorrectly (esp. if some distro
decides to do some "clever" service config switching using symlink
retargetting).

So, resolve the link manually - first to open the parent directory, and
then the final component within it. There remains a race where we
identify a link and try to read its target but it becomes a non-link
between those operations, but at least we can error out on this.