dinit-io: rework exceptions to be more lightweight
Remove extension of std::runtime_exception. Though we could with a
little trickery avoid the possibility of getting a bad_alloc when
constructing the exceptions even if they do extend runtime_exception, it
doesn't seem worth the hassle.

Just keep a simple hierarchy. For iostream_system_err, keep the errno
value as a field directly.