dinit-io: fix write_buf to take len argument
There's no point using write_buf(char *) vs write(char *) unless you
know the length of the string, since you need to compare it to the
return of write_buf in order to detect partial write. If you need to
know it already anyway, better to pass it in than have to recalculate it
(via strlen).

Also, the function is meant for writing byte sequences, which might not
be nul terminated, so a length count is needed.