dinit-io: tolerate EINTR from read/write
I suspect most OSes never return EINTR for reads of regular files
anyway, but POSIX seems to allow for it, so handle it.

While this should probably be optional if this was a general-purpose
library, within Dinit itself we'll only be using dinit-io for reading
regular files, and we don't want to abort if a signal happens to come in
at the wrong time. (If things change, this handling can be amended
then).