dinit-io: add check_buf() function to check buffer failure
In some cases it is useful to handle buffer allocation failure as an
exception (a regular out-of-memory condition, which is often easiest to
handle via an exception) while handling other errors directly.

Move io_error into the base class (from istream/ostream) and add a
check_buf() function which can be used after a failed open(...) call to
throw a std::bad_alloc if the open failed due to buffer allocation
failure.