open_with_dir: make sure to open parent_fd
A previous optimisation avoided opening the parent directory (in each
iteration of resolution) if the path was empty. However, this
potentially returned the passed-in resolve_fd as the parent directory
descriptor, and closing it would have the unintended consequence of also
closing that descriptor.

So, always make sure to dup the descriptor before returning it (or if it
is AT_FDCWD, open the current directory and return that as the parent
directory fd).