Consistently quote/tokenize user-specified values in configure script
Commands and arguments specified by the user, such as the CXX and
CXXFLAGS values, need to have standard shell tokenization performed and
so need to be used in a shell "eval" context.

In contrast, the prefix directory etc need to be shell-quoted before
being written to mconfig.

All values need to be makefile-quoted before being writtent to mconfig.

Fixes #471