try_both_argument: make the out-parameters optional
Allow dash to be specified instead of out-parameters. It's common to
check a flag for both compiling and linking, but only add the
flag to one of the compiler or link flags variables to avoid duplicate
flags. It was done by specifying a "dummy" variable, which the caller
would then unset.