Install: make '/usr/bin' default for binaries except shutdown
Make '/usr/bin' the default for dinit binaries other than shutdown (and
alias) binaries. This is what distributions are using anyway, and is
more friendly to using dinit as a regular user.

For configure-generated configuration, default sbindir will now be
'/usr/sbin' rather than '/sbin' but an alternative can easily be
suggested as before (--sbindir=xxx).