enable_disable_service: set service_file_path before issuing error
In "dinitctl enable" handling (enable_disable_service), if the service
description file for the "from" service cannot be read then an error is
issued, which includes the file path. Previously, the file path was not
been calculated at the point where the error was issued, so the
resulting error message contained an empty path; fix that by setting
service_file_path earlier.