Support "@meta enable-via" to specify default "from" for enable
The "dinitctl enable" command normally defaults to enabling "from" the
boot service (i.e. it adds the specified service as a dependent of the
boot service).

Add a "@meta" meta-command, usable in service descriptions, for
specifying metadata about a service that dinit itself does not care
about, and implement a "@meta enable-via" subcommand to inform dinitctl
which service to enable via, by default.

Documentation to follow.