dinit-service man page: capabilities: add more info
Capabilities on Linux are a badly-documented mess. Try to be a bit
clearer on exactly what this setting does - i.e. if specified, it
*replaces* the IAB vectors for service processes; if not specified, the
original IAB vectors aren't touched.

(There is no way presently to _modify_ the current vectors without
replacing them completely. The libcap library doesn't seem to provide a
straightforward means to let us do that either).