BUILD: refactor documentation about RTTI and -fno-rtti compiler option
Add more information and update documentation about RTTI and -fno-rtti,
and also a new section about using -fno-rtti on certain systems using
Clang and the Libcxxrt C++ runtime library. The documentation was a
little outdated and unclear. For instance, we have a confirmed case of
a Dinit crash on FreeBSD with the newest Clang+libcxxrt when using
-fno-rtti (#486), but the doc said

    "(and historically FreeBSD, IIRC)"

had the exception handling problem with -fno-rtti.