mconfig.Linux.sh: remove "command -v" when finding script directory
Remove "command -v "$0"" from the fragment used to find the script's
containing directory (retain use of "realpath").

The intention of this was to resolve the script name even if it was
found on the PATH, but it turns out the shell (at least bash/dash) put
the full path to the script in $0 in this case anyway, and also "command
-v $0" fails to find the script if it is not marked executable (noticed
by mobin-2008).