get_service_description_dir: validate reply before return
Check that the service description directory returned by the daemon is
not an empty string and flag a protocol error if so, in the
'get_service_description_dir' function itself, rather than in callers.

The 'get_service_description_dir' function gets the service description
directory for a service (i.e. the directory in which the service
description file was found). It should never return an empty string.