configure: always recompile test source when testing link argument
Recompile the test source before testing whether a link argument is
supported (try_ld_argument), otherwise it is potentially attempting to
link an object compiled with arbitrary flags via a previous test.

This fixes an issue where, if "-flto" was accepted at compile time but
not for linking, detection of "-lcap" would fail even if libcap is
present.