configure: fix broken cross-compilation detection
Assume the build is native if the compiler for the host and the build
machine are the same.

For example, cbuild (Chimera Linux package builder) defines both the
host and the build compiler, regardless of native or cross build.

Also, that assumption in this case is consistent with GNU autotools
behaviour.