configure: improve cross-compilation detection
Assume the build is native if the compiler for the host and the build
machine are the same and both use the same flags.

It is fairly common to set both variables to the same compiler with
same flags in a native build. For example, cbuild (Chimera Linux
package builder) defines both the host and the build compiler and
their respective compiler flags, regardless of native or cross build.

Also, that assumption in this case is consistent with GNU autotools
behaviour.